application/requirements.txt
Django ==4.0
django-import-export == 2.7.0
python-dateutil==2.8.1
django-crispy-forms==1.9.0
django-user-accounts==3.0.0
django-select2==7.1.0